Added delete function, scoring etc

+++ Backend/server/server.js
@@ -13,14 +13,27 @@ const client = new Client({
     database : "mainconnection"
 })
 
-app.use(cors())
 app.use(express.json())
+app.use(cors())
 
 app.listen(port, () => console.log(`server has started on port: ${port}`))
 
-app.get('/team', async (req, res) => {
-    const data = await Send_Players_Team()
-    res.status(200).json({data : (data)})
+app.post('/team', async (req, res) => {
+    const {package} =  await req.body
+    const user_id = await Get_User_Id(package)
+    const data = await Send_Players_Team(user_id)
+    console.log(data)
+    res.json({data : data})
+})
+
+app.post('/delete', async(req, res) => {
+    const {package} = await req.body
+    if (!package){
+        return res.status(400).send({status : 'failed'})
+    }
+    res.status(200).send({status: 'recieved'})
+    let userId = await Get_User_Id(package)
+    await client.query(`DELETE FROM userteam USING playersmain WHERE userteam.player_id = playersmain.id AND player_id = ${package.player_id} AND user_id = '${userId}';`)
 })
  
 app.post('/', (req, res) => {
@@ -58,7 +71,11 @@ app.post('/getid', (req, res) => {
         return res.status(400).send({status : 'failed'})
     }
     res.status(200).send({status: 'recieved'})
-    console.log(package.sub)
+})
+
+app.get('/teamscore', async (req, res) => {
+    const data = await calc_score()
+    res.status(200).json({data : data})
 })
 
 client.connect()
@@ -71,32 +88,34 @@ function Get_User_Id(package){
         userId = userId.replace("apple|", "")
     }else if(userId.includes("auth0|")){
         userId = userId.replace("auth0|", "")
+    }else{
+        return
     }
     return userId
 }
 
 async function Send_Team_Info(package){
     try{
-        let response = await client.query(`INSERT INTO userteam(player_id, user_id) VALUES(${Number(package.playerId)}, ${(Get_User_Id(package))});`)
+        let response = await client.query(`INSERT INTO userteam(player_id, user_id) VALUES(${Number(package.playerId)}, '${(Get_User_Id(package))}');`)
 
         let res = await client.query(`SELECT DISTINCT * FROM userteam INNER JOIN playersmain ON player_id = id WHERE user_id IN(SELECT user_id FROM(SELECT user_id, ROW_NUMBER() OVER(PARTITION BY user_id) AS occurrences FROM userteam) AS duplicates WHERE duplicates.occurrences > 4 AND position = 'Defender');`)
         if(res.rowCount > 4){
-            await client.query(`DELETE FROM userteam USING playersmain WHERE userteam.player_id = playersmain.id AND player_id = ${res.rows[0].player_id} AND position = 'Defender';`)
+            await client.query(`DELETE FROM userteam USING playersmain WHERE userteam.player_id = playersmain.id AND player_id = ${res.rows[0].player_id} AND position = 'Defender' AND user_id = '${Get_User_Id(package)}';`)
         }
 
         let res1 = await client.query(`SELECT DISTINCT * FROM userteam INNER JOIN playersmain ON player_id = id WHERE user_id IN(SELECT user_id FROM(SELECT user_id, ROW_NUMBER() OVER(PARTITION BY user_id) AS occurrences FROM userteam) AS duplicates WHERE duplicates.occurrences > 1 AND position = 'Goalkeeper');`)
         if(res1.rowCount > 1){
-            await client.query(`DELETE FROM userteam USING playersmain WHERE userteam.player_id = playersmain.id AND player_id = ${res1.rows[0].player_id} AND position = 'Goalkeeper';`)
+            await client.query(`DELETE FROM userteam USING playersmain WHERE userteam.player_id = playersmain.id AND player_id = ${res1.rows[0].player_id} AND position = 'Goalkeeper' AND user_id = '${Get_User_Id(package)}';`)
         }
 
         let res2 = await client.query(`SELECT DISTINCT * FROM userteam INNER JOIN playersmain ON player_id = id WHERE user_id IN(SELECT user_id FROM(SELECT user_id, ROW_NUMBER() OVER(PARTITION BY user_id) AS occurrences FROM userteam) AS duplicates WHERE duplicates.occurrences > 3 AND position = 'Midfielder');`)
         if(res2.rowCount > 3){
-            await client.query(`DELETE FROM userteam USING playersmain WHERE userteam.player_id = playersmain.id AND player_id = ${res2.rows[0].player_id} AND position = 'Midfielder';`)
+            await client.query(`DELETE FROM userteam USING playersmain WHERE userteam.player_id = playersmain.id AND player_id = ${res2.rows[0].player_id} AND position = 'Midfielder' AND user_id = '${Get_User_Id(package)}';`)
         }
         
         let res3 = await client.query(`SELECT DISTINCT * FROM userteam INNER JOIN playersmain ON player_id = id WHERE user_id IN(SELECT user_id FROM(SELECT user_id, ROW_NUMBER() OVER(PARTITION BY user_id) AS occurrences FROM userteam) AS duplicates WHERE duplicates.occurrences > 3 AND position = 'Attacker');`)
         if(res3.rowCount > 3){
-            await client.query(`DELETE FROM userteam USING playersmain WHERE userteam.player_id = playersmain.id AND player_id = ${res3.rows[0].player_id} AND position = 'Attacker';`)
+            await client.query(`DELETE FROM userteam USING playersmain WHERE userteam.player_id = playersmain.id AND player_id = ${res3.rows[0].player_id} AND position = 'Attacker' AND user_id = '${Get_User_Id(package)}';`)
         }
 
         return response
@@ -106,9 +125,9 @@ async function Send_Team_Info(package){
     }
 }
 
-async function Send_Players_Team(){
+async function Send_Players_Team(id){
     try{
-        var response = await client.query(`SELECT playername, nationality, teamicon, rating, position FROM userteam INNER JOIN playersmain ON player_id = id WHERE user_id = '104658729806845260677';`)
+        var response = await client.query(`SELECT id, playername, nationality, teamicon, rating, position FROM userteam INNER JOIN playersmain ON player_id = id WHERE user_id = '${id}';`)
     }catch(error){
         console.error(`Error: ${error}`)
     }
@@ -125,3 +144,26 @@ async function Get_Players_From_Db(position){
     return (response.rows)
 }
 
+async function calc_score(){
+    let score = 0
+    let response = await client.query(`SELECT appearances, goals, assists, passes, conceded, yellow, yellowred, red, minutes, duelswon, penwon, penscored, penmissed, pencommited FROM userteam INNER JOIN playersmain ON player_id = id WHERE user_id = '001219.564b1b03e5434e05b8e08af147f58012.1528'`)
+    response.rows.map((player) => {
+        if(player.appearances > 0){
+            score += player.minutes / 30
+            score += 5 * player.goals
+            score += 3 * player.assists
+            score += player.passes / 30
+            score -= player.conceded / 4
+            score -= 2 * player.yellow
+            score -= 2 * player.yellowred
+            score -= 5 * player.red
+            score += player.duelswon
+            score += player.penwon
+            score -= player.penscored
+            score += player.penmissed
+            score -= player.pencommited
+        }
+    })
+    console.log(score)
+    return(Math.round(score))
+}
